+++ feature/signin/src/main/java/tht/feature/signin/auth/VerifyActivity.kt
@@ -1,4 +1,4 @@
-package tht.feature.signin
+package tht.feature.signin.auth
 
 import android.animation.AnimatorSet
 import android.animation.ObjectAnimator
@@ -17,6 +17,7 @@ import tht.core.ui.delegate.viewBinding
 import tht.core.ui.extension.repeatOnStarted
 import tht.core.ui.extension.setSoftKeyboardVisible
 import tht.core.ui.extension.showToast
+import tht.feature.signin.R
 import tht.feature.signin.databinding.ActivityVerifyBinding
 
 @AndroidEntryPoint
