+++ feature/signin/src/main/java/tht/feature/signin/auth/VerifyEditTextWatcher.kt
@@ -1,4 +1,4 @@
-package tht.feature.signin
+package tht.feature.signin.auth
 
 import android.text.Editable
 import android.text.TextWatcher
