+++ feature/signin/src/main/java/tht/feature/signin/intro/IntroActivity.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.launch
 import tht.core.ui.delegate.viewBinding
 import tht.core.ui.extension.repeatOnStarted
 import tht.core.ui.extension.showToast
-import tht.feature.signin.PhoneAuthActivity
+import tht.feature.signin.auth.PhoneAuthActivity
 import tht.feature.signin.R
 import tht.feature.signin.databinding.ActivityIntroBinding
